Fix XPath used to find fields in WSDL.
Not all fields where been found when determining if a field should be included in the request or
response. This is due to inconsistencies in the different WSDLs. Some
used CallInfo and others use callInfo. Some WSDL's dont't have appinfo
elements.

Therefore added more XPath statments to try and cater for all eventualities.