Fix issues with xslt.
Some xs:elements where not been found by the xslt. Anti recursion was
incorrectly flagging some elements as already appearing when the name of
the element appeared as a substring in the names of other elements.

E.g

The element Value appears as a substring below and would therefore not
be included.

DetailValue PropertyName