Ensure RequesterCredentials element is added to all AbstractRequestType.
All requests via the trading API require certain credential fields to be
specified. Using the WSDL these fields do not appear in the resulting
transformation into HTML. This is because the AbstractRequestType type
does not include RequesterCredentials as a member of its element.

To fix this the XSLT now detects if it transforming an
AbstractRequestType and will perform an addition transformation of the
RequesterCredentials element. In effect the XSLT injects a new field
called RequesterCredentials.